1. Added CartPage. 2. Added userOrder page 3. addd a cartTotal state in redux (foodOrderSlice) 4. added update cartTotal reducer in foodOrdelSlice
To Do(Need to have a look and try to fix these):
a. In cartpage I brought restaurantMenuList and cartItems from redux(Note restaurantMenuList is an object as i forgot to return res.data.payload but i returned only res.data in foodOrderSlice, I did not correct it now because i have to change at many places now, so foodItemList = restaurantMenuList.payload;

b. Also added getCartTotal() function in cart.jsx but wrote cartTotal updated in redux(Need to fix this shit and make a getCartTotal() and updated in redux itself)

c. lso it might require to maintain subTotal and total separately

d. Since i am passing restaurantObject as state in useNaveigate hook to FoodDisplay component from RestaurantsDisplay/RestaurantsCard, if we reload or go back and fron in page we get error as the restaurantObj is lost if we reload the FoodDisplay page.  So try to fix this(One way is passing restaurantId as url paramenter alonfside restaurantName)